#3 - people stuff in model
