+++ expo/gbe/models/remains.py
@@ -1,8 +1,4 @@
 import os
-from datetime import (
-    datetime,
-    timedelta
-)
 from expo.settings import (
     DATETIME_FORMAT,
     TIME_FORMAT,
@@ -14,19 +10,11 @@ from django.db.models import Q
 from django.core.validators import (
     MinValueValidator,
     MaxValueValidator,
-    RegexValidator,
 )
 from django.core.exceptions import (
     ValidationError,
     NON_FIELD_ERRORS,
 )
-from django.core.urlresolvers import reverse
-from django.core.mail import send_mail
-from django.template import (
-    loader,
-    Context,
-)
-from django.contrib.auth.models import User
 from itertools import chain
 from scheduler.models import (
     Schedulable,
@@ -46,100 +34,18 @@ from model_utils.managers import InheritanceManager
 from gbe.duration import Duration
 import gbe
 import pytz
-from gbe.models import Room
+from gbe.models import (
+    Conference,
+    Performer,
+    Persona,
+    Profile,
+    Room,
+    VolunteerWindow,
+)
 
-phone_regex = '(\d{3}[-\.]?\d{3}[-\.]?\d{4})'
 
 visible_bid_query = (Q(biddable_ptr__conference__status='upcoming') |
                      Q(biddable_ptr__conference__status='ongoing'))
-
-
-class Conference(models.Model):
-    conference_name = models.CharField(max_length=128)
-    conference_slug = models.SlugField()
-    status = models.CharField(choices=conference_statuses,
-                              max_length=50,
-                              default='upcoming')
-    accepting_bids = models.BooleanField(default=False)
-
-    def __unicode__(self):
-        return self.conference_name
-
-    @classmethod
-    def current_conf(cls):
-        return cls.objects.filter(status__in=('upcoming', 'ongoing')).first()
-
-    @classmethod
-    def by_slug(cls, slug):
-        try:
-            return cls.objects.get(conference_slug=slug)
-        except cls.DoesNotExist:
-            return cls.current_conf()
-
-    @classmethod
-    def all_slugs(cls):
-        return cls.objects.order_by('-accepting_bids').values_list(
-            'conference_slug', flat=True)
-
-    def windows(self):
-        return VolunteerWindow.objects.filter(day__conference=self)
-
-    class Meta:
-        verbose_name = "conference"
-        verbose_name_plural = "conferences"
-        app_label = "gbe"
-
-
-class ConferenceDay(models.Model):
-    day = models.DateField(blank=True)
-    conference = models.ForeignKey(Conference)
-
-    def __unicode__(self):
-        return date_format(self.day, "DATE_FORMAT")
-
-    class Meta:
-        ordering = ['day']
-        verbose_name = "Conference Day"
-        verbose_name_plural = "Conference Days"
-        app_label = "gbe"
-
-
-class VolunteerWindow(models.Model):
-    start = models.TimeField(blank=True)
-    end = models.TimeField(blank=True)
-    day = models.ForeignKey(ConferenceDay)
-
-    def __unicode__(self):
-        return "%s, %s to %s" % (str(self.day),
-                                 date_format(self.start, "TIME_FORMAT"),
-                                 date_format(self.end, "TIME_FORMAT"))
-
-    def start_timestamp(self):
-        return pytz.utc.localize(datetime.combine(self.day.day, self.start))
-
-    def end_timestamp(self):
-        return pytz.utc.localize(datetime.combine(self.day.day, self.end))
-
-    def check_conflict(self, start, end):
-        starttime = self.start_timestamp()
-        endtime = self.end_timestamp()
-        has_conflict = False
-
-        if start == starttime:
-            has_conflict = True
-        elif (start > starttime and
-              start < endtime):
-            has_conflict = True
-        elif (start < starttime and
-              end > starttime):
-            has_conflict = True
-        return has_conflict
-
-    class Meta:
-        ordering = ['day', 'start']
-        verbose_name = "Volunteer Window"
-        verbose_name_plural = "Volunteer Windows"
-        app_label = "gbe"
 
 
 class Biddable(models.Model):
@@ -178,542 +84,6 @@ class Biddable(models.Model):
     @property
     def is_current(self):
         return self.conference.status in ("upcoming", "current")
-
-
-class Profile(WorkerItem):
-    '''
-    The core data about any registered user of the GBE site, barring
-    the information gathered up in the User object. (which we'll
-    expose with properties, I suppose)
-    '''
-    user_object = models.OneToOneField(User)
-    display_name = models.CharField(max_length=128, blank=True)
-
-    # used for linking tickets
-    purchase_email = models.CharField(max_length=64, blank=True, default='')
-
-    # contact info - I'd like to split this out to its own object
-    # so we can do real validation
-    # but for now, let's just take what we get
-
-    address1 = models.CharField(max_length=128, blank=True)
-    address2 = models.CharField(max_length=128, blank=True)
-    city = models.CharField(max_length=128, blank=True)
-    state = models.CharField(max_length=2,
-                             choices=states_options,
-                             blank=True)
-    zip_code = models.CharField(max_length=10, blank=True)  # allow for ext ZIP
-    country = models.CharField(max_length=128, blank=True)
-    # must have = a way to contact teachers & performers on site
-    # want to have = any other primary phone that may be preferred offsite
-    phone = models.CharField(max_length=50,
-                             validators=[RegexValidator(
-                                 regex=phone_regex,
-                                 message=phone_number_format_error)])
-    best_time = models.CharField(max_length=50,
-                                 choices=best_time_to_call_options,
-                                 default='Any',
-                                 blank=True)
-    how_heard = models.TextField(blank=True)
-
-    @property
-    def review_header(self):
-        return (['Name',
-                 'Username',
-                 'Last Login',
-                 'Email',
-                 'Purchase Email',
-                 'Phone',
-                 'Action'])
-
-    @property
-    def review_summary(self):
-        return (self.display_name,
-                self.user_object.username,
-                self.user_object.last_login,
-                self.user_object.email,
-                self.purchase_email,
-                self.phone)
-
-    def bids_to_review(self):
-        reviews = []
-        missing_reviews = []
-        if 'Act Reviewers' in self.privilege_groups:
-            reviews += Act().bids_to_review.exclude(
-                bidevaluation__evaluator=self)
-        if 'Class Reviewers' in self.privilege_groups:
-            reviews += Class().bids_to_review.exclude(
-                bidevaluation__evaluator=self)
-        if 'Costume Reviewers' in self.privilege_groups:
-            reviews += Costume().bids_to_review.exclude(
-                bidevaluation__evaluator=self)
-        if 'Vendor Reviewers' in self.privilege_groups:
-            reviews += Vendor().bids_to_review.exclude(
-                bidevaluation__evaluator=self)
-        if 'Volunteer Reviewers' in self.privilege_groups:
-            reviews += Volunteer().bids_to_review.exclude(
-                bidevaluation__evaluator=self)
-        return reviews
-
-    @property
-    def contact_email(self):
-        return self.user_object.email
-
-    @property
-    def contact_phone(self):
-        return self.phone
-
-    @property
-    def address(self):
-        address_string = str(self.address1.strip() +
-                             '\n' +
-                             self.address2.strip()).strip()
-        if len(address_string) == 0:
-            return ''
-        if (len(self.city) == 0 or
-                len(self.country) == 0 or
-                len(self.state) == 0 or
-                len(self.zip_code) == 0):
-            return ''
-        return address_string + '\n' + ' '.join((self.city + ',',
-                                                 self.state,
-                                                 self.zip_code,
-                                                 self.country))
-
-    @property
-    def special_privs(self):
-        from gbe.special_privileges import special_privileges
-        privs = [special_privileges.get(group, None) for group in
-                 self.privilege_groups]
-        return filter(lambda x: x is not None, privs)
-
-    @property
-    def privilege_groups(self):
-        groups = [group.name for
-                  group in self.user_object.groups.all().order_by('name')]
-        return groups
-
-    def alerts(self, historical=False):
-        if historical:
-            return []
-        p_alerts = []
-        if (len(self.display_name.strip()) == 0 or
-                len(self.purchase_email.strip()) == 0):
-            p_alerts.append(profile_alerts['empty_profile'] %
-                            reverse('profile_update',
-                                    urlconf='gbe.urls'))
-        expo_commitments = []
-        expo_commitments += self.get_shows()
-        expo_commitments += self.is_teaching()
-        if (len(expo_commitments) > 0 and len(self.phone.strip()) == 0):
-            p_alerts.append(profile_alerts['onsite_phone'] %
-                            reverse('profile_update',
-                                    urlconf='gbe.urls'))
-        for act in self.get_acts():
-            if act.accepted == 3 and \
-               act.is_current and \
-               (len(act.get_scheduled_rehearsals()) == 0 or
-                    not act.tech.is_complete):
-                p_alerts.append(
-                    profile_alerts['schedule_rehearsal'] %
-                    (act.title,
-                     reverse('act_techinfo_edit',
-                             urlconf='gbe.urls',
-                             args=[act.id])))
-        return p_alerts
-
-    def get_costumebids(self, historical=False):
-        costumes = self.costumes.all()
-        return (c for c in costumes if c.is_current != historical)
-
-    def get_volunteerbids(self):
-        return [vbid for vbid in self.volunteering.all() if vbid.is_current]
-
-    def get_performers(self):
-        performers = self.get_personae()
-        performers += self.get_troupes()
-        performers += self.get_combos()
-        return performers
-
-    def get_personae(self):
-        solos = self.personae.all()
-        performers = list(solos)
-        return performers
-
-    def get_troupes(self):
-        solos = self.personae.all()
-        performers = list()
-        for solo in solos:
-            performers += solo.troupes.all()
-        performers += Troupe.objects.filter(contact=self)
-        perf_set = set(performers)
-        return perf_set
-
-    def get_combos(self):
-        solos = self.personae.all()
-        performers = list()
-        for solo in solos:
-            performers += solo.combos.all()
-        performers += Combo.objects.filter(contact=self)
-        perf_set = set(performers)
-        return perf_set
-
-    def get_acts(self, show_historical=False):
-        acts = []
-        performers = self.get_performers()
-        for performer in performers:
-            acts += performer.acts.all()
-        if show_historical:
-            f = lambda a: not a.is_current
-        else:
-            f = lambda a: a.is_current
-        return filter(f, acts)
-
-    def get_shows(self):
-        acts = self.get_acts()
-        shows = [Show.objects.filter(
-            scheduler_events__resources_allocated__resource__actresource___item=act)
-            for act in acts if act.accepted == 3 and act.is_current]
-        return sum([list(s) for s in shows], [])
-
-    def get_schedule(self, conference=None):
-        '''
-        Gets all schedule items for a conference, if a conference is provided
-        Otherwise, it's the same logic as schedule() below.
-        '''
-        events = self.schedule
-        if conference:
-            conf_events = filter(
-                lambda x: x.eventitem.get_conference() == conference, events)
-        else:
-            conf_events = events
-        return conf_events
-
-    @property
-    def schedule(self):
-        '''
-        Gets all of a person's schedule.  Every way the actual human could be
-        committed:
-        - via profile
-        - via performer(s)
-        - via performing in acts
-        Returns schedule as a list of Scheduler.Events
-        NOTE:  Things that haven't been booked with start times won't be here.
-        '''
-        from scheduler.models import Event as sEvent
-        acts = self.get_acts()
-        events = sum([list(sEvent.objects.filter(
-            resources_allocated__resource__actresource___item=act))
-            for act in acts if act.accepted == 3], [])
-        for performer in self.get_performers():
-            events += [e for e in sEvent.objects.filter(
-                resources_allocated__resource__worker___item=performer)]
-        events += [e for e in sEvent.objects.filter(
-            resources_allocated__resource__worker___item=self)]
-        return sorted(set(events), key=lambda event: event.start_time)
-
-    def volunteer_schedule(self, conference=None):
-        conference = conference or Conference.current_conf()
-        return self.workeritem.get_bookings(role="Volunteer",
-                                            conference=conference).order_by(
-                                                'starttime')
-
-    def get_roles(self, conference):
-        '''
-        Gets all of a person's roles for a conference
-        '''
-        roles = get_roles_from_scheduler(
-            self.get_performers() + [self],
-            conference)
-        if self.get_shows():
-            roles += ["Performer"]
-        return roles
-
-    def get_badge_name(self):
-        badge_name = self.display_name
-        if len(badge_name) == 0:
-            badge_name = self.user_object.first_name
-        return badge_name
-
-    def is_teaching(self, historical=False):
-        '''
-        return a list of classes this user is teaching
-        '''
-        if historical:
-            return [c for c in self.workeritem.get_bookings('Teacher')
-                    if not c.is_current]
-        else:
-            return [c for c in self.workeritem.get_bookings('Teacher')
-                    if c.is_current]
-
-    def vendors(self, historical=False):
-        vendors = Vendor.objects.filter(profile=self)
-        if historical:
-            f = lambda v: not v.is_current
-        else:
-            f = lambda v: v.is_current
-        return filter(f, vendors)
-
-    def proposed_classes(self, historical=False):
-        classes = sum([list(teacher.is_teaching.all())
-                       for teacher in self.personae.all()], [])
-        if historical:
-            f = lambda c: not c.is_current
-        else:
-            f = lambda c: c.is_current
-        classes = filter(f, classes)
-        return classes
-
-    def sched_payload(self):
-        return {'name': self.display_name}
-
-    def has_role_in_event(self, role, event):
-        '''
-        Returns True if this person has the
-        given role in the given event
-        '''
-        doing_it = False
-        if role == "Performer":
-            for show in self.get_shows():
-                if show.pk == event.pk:
-                    doing_it = True
-        elif not doing_it:
-            performers = self.get_performers()
-            for person in event.roles([role]):
-                if self.pk == person._item.pk:
-                    doing_it = True
-                else:
-                    for perf in performers:
-                        if perf.pk == person._item.pk:
-                            doing_it = True
-        return doing_it
-
-    def notify_volunteer_schedule_change(self):
-        subject = "A change has been made to your Volunteer Schedule!"
-        message = loader.get_template('gbe/volunteer_schedule_update.tmpl')
-        c = Context({'profile': self})
-        if not settings.DEBUG:
-            mail_to_user(subject, message.render(c), self.user_object)
-
-    def __str__(self):
-        return self.display_name
-
-    @property
-    def describe(self):
-        return self.display_name
-
-    def __unicode__(self):
-        return self.display_name
-
-    class Meta:
-        ordering = ['display_name']
-        app_label = "gbe"
-
-
-class Performer (WorkerItem):
-    '''
-    Abstract base class for any solo, group, or troupe - anything that
-    can appear in a show lineup or teach a class
-    The fields are named as we would name them for a single performer.
-    In all cases, when applied to an aggregate (group or troup) they
-    apply to the aggregate as a whole. The Boston Baby Dolls DO NOT
-    list awards won by members of the troupe, only those won by the
-    troup. (individuals can list their own, and these can roll up if
-    we want). Likewise, the bio of the Baby Dolls is the bio of the
-    company, not of the members, and so forth.
-    '''
-    objects = InheritanceManager()
-    contact = models.ForeignKey(Profile, related_name='contact')
-    name = models.CharField(max_length=100,     # How this Performer is listed
-                            unique=True)        # in a playbill.
-    homepage = models.URLField(blank=True)
-    bio = models.TextField()
-    experience = models.PositiveIntegerField()       # in years
-    awards = models.TextField(blank=True)
-    promo_image = models.FileField(upload_to="uploads/images",
-                                   blank=True)
-    festivals = models.TextField(blank=True)     # placeholder only
-
-    def append_alerts(self, alerts):
-        '''
-        Find any alerts generated by this object's data and append them
-        to the alerts dict presented as a parameter
-        '''
-        return alerts
-
-    def get_promo_image_sizes(self, size):
-        '''
-        gets the URL for images that have been sized specifically to
-        suit GBE layouts
-            mini = the size for show/act and fashion fair grids - height = 125p
-            thumb = the size for the /gbe page -  height 75-100p
-        There is a timebased script that minifies uploaded images in these
-        two formats
-        '''
-        resized = None
-        if self.promo_image:
-            pieces = self.promo_image.name.split('/')
-            pieces.insert(-1, size)
-            resized = '/'.join(pieces)
-        return resized
-
-    @property
-    def promo_mini(self):
-        return self.get_promo_image_sizes('mini')
-
-    @property
-    def promo_thumb(self):
-        return self.get_promo_image_sizes('thumb')
-
-    def get_schedule(self):
-        return Performer.objects.get_subclass(
-            resourceitem_id=self.resourceitem_id).get_schedule()
-
-    def get_profiles(self):
-        '''
-        Gets all of the people performing in the act
-        '''
-        return Performer.objects.get_subclass(
-            resourceitem_id=self.resourceitem_id).get_profiles()
-
-    @property
-    def contact_email(self):
-        return self.contact.user_object.email
-
-    @property
-    def contact_phone(self):
-        return self.contact.phone
-
-    @property
-    def complete(self):
-        return True
-
-    @property
-    def describe(self):
-        return self.name
-
-    def __unicode__(self):
-        return self.name
-
-    class Meta:
-        ordering = ['name']
-        app_label = "gbe"
-
-
-class Persona (Performer):
-    '''
-    The public persona of one person who performs or teaches.
-    May be aggregated into a group or a troupe,
-    or perform solo, or both. A single person might conceivably maintain two
-    distinct performance identities and therefore have multiple
-    Persona objects associated with their profile. For example, a
-    person who dances under one name and teaches under another would
-    have multiple Personae.
-    performer_profile is the profile of the user who dons this persona.
-    '''
-    performer_profile = models.ForeignKey(Profile, related_name="personae")
-
-    '''
-    Returns the single profile associated with this persona
-    '''
-    def get_profiles(self):
-        return [self.performer_profile]
-
-    def get_schedule(self):
-        return self.performer_profile.get_schedule()
-
-    @property
-    def complete(self):
-        return (self.performer_profile.complete and
-                self.name is not '' and
-                self.bio is not '')
-
-    def append_alerts(self, alerts):
-        '''
-        Find any alerts generated by this object's data and append them to
-        the alerts dict presented as a parameter
-        '''
-        alerts = super(Persona, self).append_alerts(alerts)
-        return alerts
-
-    class Meta:
-        verbose_name_plural = 'personae'
-        app_label = "gbe"
-
-
-class Troupe(Performer):
-    '''
-    Two or more performers working together as an established entity. A troupe
-    connotes an entity with a stable membership, a history, and hopefully a
-    future. This suggests that a troupe should have some sort of legal
-    existence, though this is not required for GBE. Further specification
-    welcomed.
-    Troupes are distinct from Combos in their semantics, but at this time they
-    share the same behavior.
-    '''
-    membership = models.ManyToManyField(Persona,
-                                        related_name='troupes')
-
-    '''
-        Gets all of the people performing in the act.
-        For troupe, that is every profile of every member in membership
-    '''
-    def get_profiles(self):
-        profiles = []
-        for member in Persona.objects.filter(troupes=self):
-            profiles += member.get_profiles()
-        return profiles
-
-    def get_schedule(self):
-        return []
-
-    def append_alerts(self, alerts):
-        '''
-        Find any alerts generated by this object's data and append them
-        to the alerts dict presented as a parameter
-        '''
-        alerts = super(Troupe, self).append_alerts()
-        return alerts
-
-    class Meta:
-        app_label = "gbe"
-
-
-class Combo (Performer):
-    '''
-    Two or more performers (Personae), working together, on a temporary
-    or ad-hoc basis. For example, two performers who put together a
-    routine for the GBE but do not otherwise perform together would be
-    a Combo and not a Troupe. The distinction between Combo and Troupe
-    is basically semantic, and the separation is intended to aid in
-    maintaining that semantic distinction. If it is inconvenient, the
-    separation need not persist in the code.
-    '''
-    membership = models.ManyToManyField(Persona,
-                                        related_name='combos')
-    '''
-        Gets all of the people performing in the act.
-        For  combo, that is every profile of every member in membership
-    '''
-    def get_profiles(self):
-        profiles = []
-        for member in Persona.objects.filter(combos=self):
-            profiles += member.get_profiles()
-        return profiles
-
-    def get_schedule(self):
-        return []
-
-    def append_alerts(self, alerts):
-        '''
-        Find any alerts generated by this object's data and append them
-        to the alerts dict presented as a parameter
-        '''
-        alerts = super(Combo, self).append_alerts()
-        return alerts
-
-    class Meta:
-        app_label = "gbe"
 
 
 ###################
@@ -1772,7 +1142,3 @@ class Costume(Biddable):
 
     class Meta:
         app_label = "gbe"
-
-
-def mail_to_user(subject, message, user):
-    send_mail(subject, message, settings.DEFAULT_FROM_EMAIL, [user.email])
